Correct the callback of the HS_D_OC pin
In the YV35, the signal of the MP5990 D_OC pin is transferred to the
GPIOG2, which registers the ISR_MB_THROTTLE callback. This commit uses
the YV35 config to correct the same GPIO pin in the MFZ.

The ISR_HSC_OC callback is assigned to the GPIOC0 in YV35, and it is not
clear how the MFZ implements this hardware design. And thus, the
ISR_HSC_OC callback is temporarily unused in the MFZ.

Signed-off-by: Scron Chang <Scron.Chang@quantatw.com>